+++ docs/apidoc.md
@@ -1,6 +1,4 @@
-
-## BaseColum
-::: sweetvalidation.models
+::: sweet_validation.models
     options:
       heading_level: 3
