+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "sweetvalidation"
+name = "sweet_validation"
 version = "0.1.0"
 description = "Validate data given schemes in a sweet way"
 readme = "README.md"
